Update how mutation works
The mutation was weird previously because I had to pass the compile time
size of the chromosome into the mutation, and the mutator function had
to be aware of how to index into the type whether it was an aggregate or
tuple like.

In this new formulation, the mutator functions now just take a reference
to a parameter and mutates it. The mutation strategy is now in charge
of whether the mutation happens, as well as indexing into the chromosome
type. This simplification will lend itself to future mutation strategies
that need to do chromosome repair after a mutation happens in one of the
chromosome members.

The ConditionallyApplyFunctions utility is a little hard to read if one
is not familiar with TMP, but it makes it so the independent mutator can
do a "compile time ranged for loop" which allows for better compile time
checking that the mutator functions work with the chromosome.